DataFormContainerWidget converts "checkbox" values to boolean
`new FormData(formRef.current).entries()` will only have a value for a checkbox, if it is actually set. If it is not set, _not_ key will be used. Second the value with be the string `"on"` instead of the boolean `true`. With this code, a input checkbox is always present in `attributes` and always has a boolean value (`true` or `false`)