Rerender <source> elements on change
Fixes #767.

As far as I can see, if one changes the `src` of the `<source>` dom element, the browser (at least Chromium) will ignore this `<source>` and pretend that it does not exist.

With the explicit `key` we tell React to recreate a new `<video>` dom element, if the content URL or it type changes.

It is also a better solution for #4767 (in Scrivito)