Use react-compiler
React compiler shipped as version 1.0 earlier this month (see [1]). It optimizes rendering speed by automatically memorizing components, so that they rerender less often.

At least of now there is no rust-based (either SWC or oxc) version of the react-compiler available (see [2]). That is why this commit changes from `@vitejs/plugin-react-swc` to `@vitejs/plugin-react`, in order to be able to use the recommended `babel-plugin-react-compiler`.

[1] https://react.dev/blog/2025/10/07/react-compiler-1
[2] https://github.com/oxc-project/oxc/issues/10048